+++ cli/js.rs
@@ -25,11 +25,10 @@ fn cli_snapshot() {
     deno_core::StartupData::Snapshot(CLI_SNAPSHOT),
     false,
   );
-  deno_core::js_check(isolate.execute("<anon>", "bootstrapMainRuntime()"));
   deno_core::js_check(isolate.execute(
     "<anon>",
     r#"
-      if (!window) {
+      if (!(bootstrapMainRuntime && bootstrapWorkerRuntime)) {
         throw Error("bad");
       }
       console.log("we have console.log!!!");
@@ -46,7 +45,7 @@ fn compiler_snapshot() {
   deno_core::js_check(isolate.execute(
     "<anon>",
     r#"
-      if (!bootstrapTsCompiler) {
+    if (!(bootstrapTsCompilerRuntime && bootstrapTsCompilerRuntime)) {
         throw Error("bad");
       }
       console.log(`ts version: ${ts.version}`);
