+++ cli/js/lib.deno_main.d.ts
@@ -36,7 +36,6 @@ declare interface Window {
   performance: __performanceUtil.Performance;
   onmessage: (e: { data: any }) => void;
   onerror: undefined | typeof onerror;
-  bootstrapWorkerRuntime: typeof __workerMain.bootstrapWorkerRuntime;
   workerClose: typeof __workerMain.workerClose;
   postMessage: typeof __workerMain.postMessage;
   Worker: typeof __workers.WorkerImpl;
@@ -95,7 +94,6 @@ declare let onerror:
       e: Event
     ) => boolean | void)
   | undefined;
-declare const bootstrapWorkerRuntime: typeof __workerMain.bootstrapWorkerRuntime;
 declare const workerClose: typeof __workerMain.workerClose;
 declare const postMessage: typeof __workerMain.postMessage;
 declare const Worker: typeof __workers.WorkerImpl;
@@ -1525,9 +1523,7 @@ declare namespace __workerMain {
   export let onmessage: (e: { data: any }) => void;
   export function postMessage(data: any): void;
   export function getMessage(): Promise<any>;
-  export let isClosing: boolean;
   export function workerClose(): void;
-  export function bootstrapWorkerRuntime(): Promise<void>;
 }
 
 declare namespace __workers {
